finish plumbing, abstract out method for collecting stats
