+++ src/adapter/src/explain/optimizer_trace.rs
@@ -134,6 +134,7 @@ impl OptimizerTrace {
         config: &ExplainConfig,
         features: &OptimizerFeatures,
         humanizer: &dyn ExprHumanizer,
+        cardinality_stats: &StatisticsOracle,
         row_set_finishing: Option<RowSetFinishing>,
         target_cluster: Option<&Cluster>,
         dataflow_metainfo: DataflowMetainfo,
@@ -147,6 +148,7 @@ impl OptimizerTrace {
                 config,
                 features,
                 humanizer,
+                cardinality_stats,
                 row_set_finishing.clone(),
                 target_cluster.map(|c| c.name.as_str()),
                 dataflow_metainfo.clone(),
@@ -303,6 +305,7 @@ impl OptimizerTrace {
         self,
         features: &OptimizerFeatures,
         humanizer: &dyn ExprHumanizer,
+        cardinality_stats: &StatisticsOracle,
         row_set_finishing: Option<RowSetFinishing>,
         target_cluster: Option<&Cluster>,
         dataflow_metainfo: DataflowMetainfo,
@@ -314,6 +317,7 @@ impl OptimizerTrace {
                 &ExplainConfig::default(),
                 features,
                 humanizer,
+                cardinality_stats,
                 row_set_finishing,
                 target_cluster,
                 dataflow_metainfo,
@@ -337,6 +341,7 @@ impl OptimizerTrace {
         config: &ExplainConfig,
         features: &OptimizerFeatures,
         humanizer: &dyn ExprHumanizer,
+        cardinality_stats: &StatisticsOracle,
         row_set_finishing: Option<RowSetFinishing>,
         target_cluster: Option<&str>,
         dataflow_metainfo: DataflowMetainfo,
@@ -353,7 +358,7 @@ impl OptimizerTrace {
             config,
             features,
             humanizer,
-            cardinality_stats: &StatisticsOracle::default(), // empty stats // !!!(mgree) implement
+            cardinality_stats,
             used_indexes: Default::default(),
             finishing: row_set_finishing.clone(),
             duration: Default::default(),
@@ -376,7 +381,7 @@ impl OptimizerTrace {
             config,
             features,
             humanizer,
-            cardinality_stats: &StatisticsOracle::default(), // empty stats // !!!(mgree) implement
+            cardinality_stats,
             used_indexes: Default::default(),
             finishing: row_set_finishing,
             duration: Default::default(),
